Revert "Rename files and reports funcs"
